[ALPS04056212] aee: fix build error
format %x expects argument of type unsigned int,
but argument 4 has type long unsigned int [-Werror=format=]

MTK-Commit-Id: 7f5ee58c56e83cc0d7263d5b3ca4296e82824889

CR-Id: ALPS04056212
Feature: Android Exception Engine(AEE)

+++ drivers/misc/mediatek/aee/mrdump/mrdump_mini.c
@@ -787,7 +787,7 @@ static void mrdump_mini_build_elf_misc(void)
 	} else {
 		LOGE("minirdump: unexpected addr:0x%x, size:0x%x(0x%x)\n",
 			mrdump_mini_addr, mrdump_mini_size,
-			MRDUMP_MINI_HEADER_SIZE);
+			(unsigned int)MRDUMP_MINI_HEADER_SIZE);
 		mrdump_mini_fatal("illegal addr size");
 	}
 	mrdump_mini_add_misc_pa(task_info_va, task_info_pa,
